+++ ttnn/cpp/ttnn/distributed/distributed_pybind.cpp
@@ -11,7 +11,6 @@
 #include <tt-metalium/command_queue.hpp>
 #include "tt-metalium/mesh_coord.hpp"
 #include "distributed_tensor.hpp"
-#include "distributed_tensor.cpp"
 #include "ttnn/distributed/api.hpp"
 #include "ttnn/distributed/distributed_tensor_config.hpp"
 #include "ttnn/tensor/tensor_utils.hpp"
@@ -346,8 +345,8 @@ void py_module(py::module& module) {
                back to all SubDevice IDs.
            )doc");
 
-    auto py_tensor_to_mesh =
-        static_cast<py::class_<TensorToMesh, std::unique_ptr<TensorToMesh>>>(module.attr("TensorToMesh"));
+    auto py_tensor_to_mesh = static_cast<py::class_<TensorToMesh, ConcreteTensorToMesh, std::unique_ptr<TensorToMesh>>>(
+        module.attr("TensorToMesh"));
     py_tensor_to_mesh
         .def(py::init([]() -> std::unique_ptr<TensorToMesh> { return std::make_unique<ConcreteTensorToMesh>(); }))
         .def("map", &TensorToMesh::map)
@@ -359,13 +358,13 @@ void py_module(py::module& module) {
 
     py_replicate_tensor_to_mesh
         .def(
-            py::init([](MeshDevice& mesh_device) -> std::unique_ptr<TensorToMesh> {
-                return ttnn::distributed::replicate_tensor_to_mesh_mapper(mesh_device);
+            py::init([](MeshDevice& mesh_device) -> std::unique_ptr<ReplicateTensorToMesh> {
+                return std::make_unique<ReplicateTensorToMesh>(ReplicateTensorToMesh(mesh_device.num_devices()));
             }),
             py::kw_only(),
             py::arg("mesh_device"))
         .def(
-            py::init([](size_t num_devices) -> std::unique_ptr<TensorToMesh> {
+            py::init([](size_t num_devices) -> std::unique_ptr<ReplicateTensorToMesh> {
                 return std::make_unique<ReplicateTensorToMesh>(ReplicateTensorToMesh(num_devices));
             }),
             py::kw_only(),
@@ -380,14 +379,14 @@ void py_module(py::module& module) {
         module.attr("ShardTensorToMesh"));
     py_shard_tensor_to_mesh
         .def(
-            py::init([](MeshDevice& mesh_device, int dim) -> std::unique_ptr<TensorToMesh> {
-                return ttnn::distributed::shard_tensor_to_mesh_mapper(mesh_device, dim);
+            py::init([](MeshDevice& mesh_device, int dim) -> std::unique_ptr<ShardTensorToMesh> {
+                return std::make_unique<ShardTensorToMesh>(ShardTensorToMesh(mesh_device, dim));
             }),
             py::kw_only(),
             py::arg("mesh_device"),
             py::arg("dim"))
         .def(
-            py::init([](size_t num_devices, int dim) -> std::unique_ptr<TensorToMesh> {
+            py::init([](size_t num_devices, int dim) -> std::unique_ptr<ShardTensorToMesh> {
                 return std::make_unique<ShardTensorToMesh>(ShardTensorToMesh(num_devices, dim));
             }),
             py::kw_only(),
@@ -407,17 +406,18 @@ void py_module(py::module& module) {
             py::init(
                 [](MeshDevice& mesh_device,
                    const MeshShape& mesh_shape,
-                   const Shard2dConfig& config) -> std::unique_ptr<TensorToMesh> {
-                    return ttnn::distributed::shard_tensor_to_2d_mesh_mapper(mesh_device, mesh_shape, config);
+                   const Shard2dConfig& config) -> std::unique_ptr<ShardTensorTo2dMesh> {
+                    return std::make_unique<ShardTensorTo2dMesh>(ShardTensorTo2dMesh(mesh_device, mesh_shape, config));
                 }),
             py::kw_only(),
             py::arg("mesh_device"),
             py::arg("mesh_shape"),
             py::arg("config"))
         .def(
-            py::init([](const MeshShape& mesh_shape, const Shard2dConfig& config) -> std::unique_ptr<TensorToMesh> {
-                return std::make_unique<ShardTensorTo2dMesh>(ShardTensorTo2dMesh(mesh_shape, config));
-            }),
+            py::init(
+                [](const MeshShape& mesh_shape, const Shard2dConfig& config) -> std::unique_ptr<ShardTensorTo2dMesh> {
+                    return std::make_unique<ShardTensorTo2dMesh>(ShardTensorTo2dMesh(mesh_shape, config));
+                }),
             py::kw_only(),
             py::arg("mesh_shape"),
             py::arg("config"))
@@ -427,8 +427,8 @@ void py_module(py::module& module) {
             py::arg("tensor"))
         .def("config", &ShardTensorTo2dMesh::config);
 
-    auto py_mesh_to_tensor =
-        static_cast<py::class_<MeshToTensor, std::unique_ptr<MeshToTensor>>>(module.attr("MeshToTensor"));
+    auto py_mesh_to_tensor = static_cast<py::class_<MeshToTensor, ConcreteMeshToTensor, std::unique_ptr<MeshToTensor>>>(
+        module.attr("MeshToTensor"));
     py_mesh_to_tensor
         .def(py::init([]() -> std::unique_ptr<MeshToTensor> { return std::make_unique<ConcreteMeshToTensor>(); }))
         .def("compose", &MeshToTensor::compose);
@@ -437,8 +437,8 @@ void py_module(py::module& module) {
         module.attr("ConcatMeshToTensor"));
     py_concat_mesh_to_tensor
         .def(
-            py::init([](int dim) -> std::unique_ptr<MeshToTensor> {
-                return ttnn::distributed::concat_mesh_to_tensor_composer(dim);
+            py::init([](int dim) -> std::unique_ptr<ConcatMeshToTensor> {
+                return std::make_unique<ConcatMeshToTensor>(dim);
             }),
             py::kw_only(),
             py::arg("dim"))
@@ -452,9 +452,15 @@ void py_module(py::module& module) {
             module.attr("Concat2dMeshToTensor"));
     py_concat_2d_mesh_to_tensor
         .def(
-            py::init([](MeshDevice& mesh_device, const Concat2dConfig& config) -> std::unique_ptr<MeshToTensor> {
-                return ttnn::distributed::concat_2d_mesh_to_tensor_composer(mesh_device, config);
-            }),
+            py::init(
+                [](MeshDevice& mesh_device, const Concat2dConfig& config) -> std::unique_ptr<Concat2dMeshToTensor> {
+                    TT_FATAL(
+                        config.row_dim != config.col_dim,
+                        "Dimensions in 'dims' must be different; got row_dim: {}, col_dim: {}",
+                        config.row_dim,
+                        config.col_dim);
+                    return std::make_unique<Concat2dMeshToTensor>(mesh_device, config);
+                }),
             py::kw_only(),
             py::arg("mesh_device"),
             py::arg("config"))
@@ -515,6 +521,40 @@ void py_module(py::module& module) {
            Tensor: The shard of the tensor corresponding to the device.
    )doc");
     module.def("get_device_tensors", &get_device_tensors, py::arg("tensor"), py::kw_only());
+    module.def(
+        "replicate_tensor_to_mesh_mapper",
+        [](MeshDevice& mesh_device) -> std::unique_ptr<TensorToMesh> {
+            return replicate_tensor_to_mesh_mapper(mesh_device);
+        },
+        py::arg("mesh_device"));
+    module.def(
+        "shard_tensor_to_mesh_mapper",
+        [](MeshDevice& mesh_device, int dim) -> std::unique_ptr<TensorToMesh> {
+            return shard_tensor_to_mesh_mapper(mesh_device, dim);
+        },
+        py::arg("mesh_device"),
+        py::arg("dim"));
+    module.def(
+        "shard_tensor_to_2d_mesh_mapper",
+        [](MeshDevice& mesh_device,
+           const MeshShape& mesh_shape,
+           const Shard2dConfig& config) -> std::unique_ptr<TensorToMesh> {
+            return shard_tensor_to_2d_mesh_mapper(mesh_device, mesh_shape, config);
+        },
+        py::arg("mesh_device"),
+        py::arg("mesh_shape"),
+        py::arg("config"));
+    module.def(
+        "concat_mesh_to_tensor_composer",
+        [](int dim) -> std::unique_ptr<MeshToTensor> { return concat_mesh_to_tensor_composer(dim); },
+        py::arg("dim"));
+    module.def(
+        "concat_2d_mesh_to_tensor_composer",
+        [](MeshDevice& mesh_device, const Concat2dConfig& config) -> std::unique_ptr<MeshToTensor> {
+            return concat_2d_mesh_to_tensor_composer(mesh_device, config);
+        },
+        py::arg("mesh_device"),
+        py::arg("config"));
     // TODO: overload this method to enable selection of a subset of shards with a config or something before passing to
     // aggregate
     module.def(
